Move speculos testcontainer setup to its own file
